pyproject.toml
[project]
name = "jarvis"
version = "0.2.0"
description = "Local voice AI assistant with MCP support"
requires-python = ">=3.12"

dependencies = [
    # Agent Framework
    "langgraph>=0.2",
    "langchain-ollama>=0.2",
    "langchain-core>=0.3",

    # MCP Integration
    "langchain-mcp-adapters>=0.1",
    "python-dotenv>=1.0",

    # CLI
    "click>=8.0",

    # Utilities
    "exa-py>=1.0",
]

[project.optional-dependencies]
voice = [
    "faster-whisper>=1.0",
    "kokoro-onnx>=0.4",
    "sounddevice>=0.4",
    "pynput>=1.7",
]
full = [
    "jarvis[voice]",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "plyer>=2.0",
    "rank_bm25>=0.2",
]

[project.scripts]
jarvis = "jarvis.__main__:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
